Initial solution to the problem
By factorising the count of airline purchase tickets' amount to
share as many users as possible the tuples.

For instance, if a user has purchased 10 airline tickets for 10000€
and another user has bought 5 airline tickets for 10000€ we can
share several tuples in the calculations as follows:

  - 10 = 1 + 2 + 3 + 4
  - 5  = 1 + 4

We are sharing the following tuples:

   (10000, 1)
   (10000, 4)

Without repeating any tuple in the user's report maximising the
number of elements to use to match as accurate as possible the
unanonymised result.